refactor : RacingCar class의 메소드 분리 <description> 1. element를 만드는 함수를 element 파일에 모아서 모듈화 하였다. 2. 자동차가 이동한 거리의 진행사항을 보여주는 함수를 progress 파일에서    모듈화하였다. 3. 우승자를 결정지어서 화면에 표시하는 함수를 winner 파일에 모아서    모듈화 하였다. 4. html에서 js파일을 불러올 때, 실제로 play하는 play.js 파일만 불러왔고,    play파일에서 모든 js파일이 로드되기 때문에 오류 없는 것을 확인하였다.
