+++ index.html
@@ -28,8 +28,6 @@
         <ul id="result-list" style="padding: 0"></ul>
       </div>
     </div>
-    <script type="module" src="src/car.js"></script>
-    <script type="module" src="src/index.js"></script>
     <script type="module" src="src/play.js"></script>
   </body>
 </html>
