+++ src/progress.js
@@ -1,0 +1,5 @@
+import {createResultList} from './element.js';
+
+export default function showOnGameProgress(carsArray) {
+  createResultList(carsArray);
+}
